// packages/Common/UI/Components/Types/FieldType.ts
enum FieldType {
  ObjectID = "ObjectID",
  Heading = "Heading",
  Name = "Name",
  File = "File",
  Percent = "Percent",
  ImageFile = "ImageFile",
  Hostname = "Hostname",
  URL = "URL",
  Phone = "Phone",
  Route = "Route",
  Number = "Number",
  Password = "Password",
  Dropdown = "Dropdown",
  MultiSelectDropdown = "MultiSelectDropdown",
  Text = "Text",
  Email = "Email",
  Date = "Date",
  DateTime = "DateTime",
  LongText = "LongText",
  Color = "Color",
  Port = "Port",
  HiddenText = "HiddenText",
  Actions = "Actions",
  Boolean = "Boolean",
  Entity = "Entity",
  EntityArray = "EntityArray",
  Markdown = "Markdown",
  HTML = "HTML",
  CSS = "CSS",
  JavaScript = "JavaScript",
  DictionaryOfStrings = "DictionaryOfStrings",
  JSON = "JSON",
  YAML = "YAML",
  USDCents = "USDCents",
  Element = "Element",
  Minutes = "Minutes",
  ArrayOfText = "ArrayOfText",
  Code = "Code",
  InlineCode = "InlineCode",
  Icon = "Icon",
}

export default FieldType;